Refactor: 새로운 Material 및 Mesh-SubMesh 구조를 사용하도록 기존 코드 반영
 - Mesh-SubMesh 구조를 기반으로 드로우 되도록 코드 변경(중간 단계)

+++ ModelViewerDx11/Renderer/Resources/RenderTypes.h
@@ -86,7 +86,7 @@ namespace renderer
         // TODO: improve - 좀 더 깔끔한 네이밍이 있을지.?
         MaterialParameter MaterialParam;
         // MEMO: 텍스처
-        HashID TextureHashes[eTextureType::TextureTypeCount];
+        HashID TextureHashes[static_cast<uint8_t>(eTextureType::TextureTypeCount)];
         // MEMO: 셰이더
         // TODO: Shader도 각 Shader마다 CB 슬롯과 상태별 Bind Slot들을 매핑해줄 무언가가 필요함.
         eShader ShaderType;
